refactor: 패키지 분리작업 진행.
