+++ src/main/java/lotto/view/Output.java
@@ -1,8 +1,8 @@
 package lotto.view;
 
-import lotto.Lotto;
-import lotto.LottoResult;
-import lotto.Lottos;
+import lotto.domain.Lotto;
+import lotto.domain.LottoResult;
+import lotto.domain.Lottos;
 
 import java.util.List;
