+++ src/test/java/lotto/domain/LottoResultTest.java
@@ -1,5 +1,9 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.Lotto;
+import lotto.domain.LottoResult;
+import lotto.domain.Lottos;
+import lotto.domain.Number;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
