+++ src/test/java/lotto/domain/MoneyTest.java
@@ -1,5 +1,6 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.Money;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
