+++ src/test/java/lotto/domain/RandomNumberStrategyTest.java
@@ -1,5 +1,7 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.LottoNumberStrategy;
+import lotto.domain.RandomNumberStrategy;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
